Contents

Contents

Index

types and methods

Index

GCIdentifier.GCPair

GCIdentifier.MolecularGraphJL

GCIdentifier.RDKitLib

GCIdentifier.get_grouplist

GCIdentifier.get_groups_from_smiles

GCIdentifier.group_replace

GCIdentifier.@gcstring_str

types and methods

GCIdentifier.GCPair — Type

GCPair

Struct used to hold a description of a group. contains the SMARTS string necessary to match the group within a SMILES query, and the assigned name.

GCIdentifier.RDKitLib — Type

RDKitLib

Struct used to select the RDKit library (via RDKitMinimalLib.jl package) to perform the group search. Default in Linux and Mac.

GCIdentifier.MolecularGraphJL — Type

MolecularGraphJL

Struct used to select the MolecularGraph.jl library to perform the group search. Default in Windows.

GCIdentifier.get_grouplist — Function

get_grouplist(x)

Should return a Vector{GCPair} containing the available groups for SMILES matching.

GCIdentifier.get_groups_from_smiles — Function

get_groups_from_smiles(smiles::String,groups,lib = DEFAULTLIB;connectivity = false)

Given a SMILES string and a group list (groups::Vector{GCPair}), returns a list of groups and their corresponding amount.

If connectivity is true, then it will additionally return a vector containing the amount of bonds between each pair.

lib allows the selection of a molecular library to perform the substructure matching. the two available options are:

RDKitLib() : uses RDKit (via the RDKitMinimalLib.jl package) to perform the substructure matching. Default in Linux and Mac.

MolecularGraphJL() : uses MolecularGraph.jlto perform the substructure matching. Default in Windows (due to a bug in RDKit in this particular Operating System).

Examples

julia> get_groups_from_smiles("CCO",UNIFACGroups)
("CCO", ["CH3" => 1, "CH2" => 1, "OH(P)" => 1])

julia> get_groups_from_smiles("CCO",JobackGroups,connectivity = true)
("CCO", ["-CH3" => 1, "-CH2-" => 1, "-OH (alcohol)" => 1], [("-CH3", "-CH2-") => 1, ("-CH2-", "-OH (alcohol)") => 1])

GCIdentifier.@gcstring_str — Macro

@gcstring_str(str)

given a string of the form "Group1:n1;Group2:2", returns ["Group1" => n1,"Group2 => n2]

GCIdentifier.group_replace — Function

group_replace(grouplist,keys...)

given a group list generated by get_groups_from_smiles, replaces certain groups in grouplist with the values specified in keys.

Examples

groups1 = get_groups_from_smiles("CCO", UNIFACGroups) #["CH3" => 1, "CH2" => 1, "OH(P)" => 1]
#we replace each "OH(P)" with 1 "OH" group
#and each "CH3" group with 3 "H" group and 1 "C" group
groups2 = group_replace(groups1[2],"OH(P)" => ("OH" => 1), "CH3" => [("C" => 1),("H" => 3)])